refactor: move A2A scope mappings to interface, fix custom prefix support
1. Move get_a2a_scope_mappings() from os/scopes.py to interfaces/a2a/scopes.py
   - A2A now owns its scope definitions (consistent with AGUI pattern)
   - Scopes only exist when A2A interface is mounted
   - Remove A2A routes from get_default_scope_mappings()

2. Fix get_resource_context_from_path() for custom A2A prefixes
   - Previously only handled /a2a/ prefix
   - Now handles any prefix (e.g., A2A(prefix="/api/v2"))
   - Per-resource scopes like agents:my-agent:run now work everywhere

3. Clean up interface code comments/docstrings per codebase patterns
   - Remove verbose docstrings from AGUI.__init__
   - Simplify comments in scope mapping methods
   - Condense authenticates_own_requests field comment
