+++ libs/agno/agno/os/interfaces/a2a/a2a.py
@@ -44,10 +44,6 @@ class A2A(BaseInterface):
         return self.router
 
     def get_scope_mappings(self) -> dict:
-        from agno.os.scopes import get_a2a_scope_mappings
+        from agno.os.interfaces.a2a.scopes import get_a2a_scope_mappings
 
-        # Pass the prefix verbatim — the router mounts self.prefix as-is, so the scope-map
-        # keys must be built from the same value. A fallback here (e.g. `or "/a2a"`) would
-        # diverge for A2A(prefix=""): routes mounted at the root, scope map keyed under
-        # /a2a, every route unmapped -> default-allow.
         return get_a2a_scope_mappings(self.prefix)
