+++ libs/agno/agno/os/interfaces/base.py
@@ -18,13 +18,8 @@ class BaseInterface(ABC):
     prefix: str
     tags: List[str]
 
-    # Whether this interface verifies the authenticity of its own inbound requests
-    # (e.g. Slack/Telegram/WhatsApp verify a signing secret in their routers). When
-    # True, AgentOS excludes the interface's route prefix from the central auth layer,
-    # since the interface authenticates callers itself. Interfaces that do NOT
-    # self-authenticate (the default) stay behind AuthMiddleware and require a valid
-    # AgentOS credential once authentication is enabled -- see
-    # AgentOS._add_auth_middleware.
+    # When True, interface handles its own auth (e.g. webhook signatures for Slack/Telegram)
+    # and is excluded from central AuthMiddleware
     authenticates_own_requests: bool = False
 
     router: APIRouter
